+++ lib/llm/agent.rb
@@ -339,6 +339,15 @@ module LLM
     end
 
     ##
+    # @param [LLM::Tracer, nil] other
+    #  A tracer, or nil.
+    # @return [void]
+    def tracer=(other)
+      @ctx.tracer = other
+      @tracer = other
+    end
+
+    ##
     # @return [LLM::Stream, #<<, nil]
     #  Returns a stream object, or nil
     def stream