+++ lib/llm/context.rb
@@ -422,6 +422,14 @@ module LLM
     end
 
     ##
+    # @param [LLM::Tracer, nil] other
+    #  A tracer, or nil.
+    # @return [void]
+    def tracer=(other)
+      @llm.tracer = nil
+    end
+
+    ##
     # @return [LLM::Stream, #<<, nil]
     #  Returns a stream object, or nil
     def stream