+++ lib/llm/provider.rb
@@ -286,7 +286,7 @@ class LLM::Provider
   #  A tracer
   # @return [void]
   def tracer=(tracer)
-    @tracer = tracer
+    @tracer = tracer || LLM::Tracer::Null.new(self)
   end
 
   ##